+++ core/src/com/geometric/wars/screens/AbstractMenuScreen.java
@@ -1,0 +1,49 @@
+package com.geometric.wars.screens;
+
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.graphics.GL20;
+import com.badlogic.gdx.graphics.OrthographicCamera;
+import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.badlogic.gdx.scenes.scene2d.Stage;
+import com.badlogic.gdx.scenes.scene2d.ui.Image;
+import com.badlogic.gdx.utils.viewport.FitViewport;
+import com.geometric.wars.GeometricWars;
+
+public abstract class AbstractMenuScreen extends AbstractScreen {
+
+    protected Stage stage;
+    protected SpriteBatch spriteBatch;
+
+    public AbstractMenuScreen(GeometricWars game) {
+        super(game);
+        stage = new Stage(new FitViewport(Gdx.graphics.getWidth(), Gdx.graphics.getHeight(), camera));
+        spriteBatch = new SpriteBatch();
+    }
+
+    @Override
+    protected void createCamera() {
+        camera = new OrthographicCamera(Gdx.graphics.getWidth(), Gdx.graphics.getHeight());
+        camera.update();
+    }
+
+    @Override
+    public void resize(int width, int height) {
+        stage.getViewport().update(width, height, true);
+    }
+
+    @Override
+    public void dispose() {
+        super.dispose();
+        spriteBatch.dispose();
+        stage.dispose();
+    }
+
+    @Override
+    public void render(float delta) {
+        Gdx.gl.glClearColor(0, 0, 0, 0);
+        Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
+        stage.act(delta);
+        stage.draw();
+    }
+}
